backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
sqlalchemy
mysql
PyMySQL
cryptography

python-dotenv
email-validator
python-multipart

# crawling
requests
beautifulsoup4

# Password encryption
passlib
bcrypt==4.0.1

pydantic[email]